fix(#21): pr 피드백 반영하기
- 회색 텍스트 색상은 Muted로 통일하기
- LotusSearchBar는 CCP 대신에 Hook을 사용하기
- set 함수를 직접 전달하지 않고 감싸서 전달하ㅣ
